requirements.txt
pathway
streamlit
openai
pandas
numpy
plotly
requests
python-dotenv
faker
colorama
watchdog